requirements.txt
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.3
Flask-DotEnv==0.1.2
Flask-JSON==0.4.0
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mongoengine==0.27.0
PyJWT==2.7.0
pymongo==4.3.3
pyrsistent==0.19.3
python-dotenv==1.0.0
Werkzeug==2.3.4
WTForms==3.0.1
